+++ Assignment_1/sim.py
