+++ Assignment_1/simConst.py
