+++ Assignment_2/robot_model.py
@@ -1,0 +1,15 @@
+"""
+Mobile robot simulation setup
+@author: Bijo Sebastian 
+"""
+
+import numpy as np
+
+def simulate(robot_state, control_input, delta_T = 0.5):
+   
+    x_new = new x coordinate of robot
+    y_new = new y coordinate of robot
+    theta_new = new orinetation of robot
+    
+    return ([x_new, y_new, theta_new])
+
