+++ Assignment_2/sim.py
