+++ Assignment_2/simConst.py
